modify the toolkit doc
