--- tools/dbfork/src/test/resources/fork.conf
@@ -1,58 +1,0 @@
-witnesses = [
-  {
-    address = "TS1hu4ZCcwBFYpQqUGoWy1GWBzamqxiT5W"
-    url = "http://meme5.com"
-    voteCount = 100000036
-  },
-  {
-    address = "TRY18iTFy6p8yhWiCt1dhd2gz2c15ungq3"
-    voteCount = 100000035
-  },
-  {
-    address = "TKmyxLsRR2FWMVEHaQA2pZh1xB7oXPXzG1"
-  }
-]
-
-accounts = [
-  {
-    address = "TS1hu4ZCcwBFYpQqUGoWy1GWBzamqxiT5W"
-    accountName = "Meme"
-    balance = 99000000000000000
-  },
-  {
-    address = "TRY18iTFy6p8yhWiCt1dhd2gz2c15ungq3"
-    accountType = "Normal"
-    balance = 99000000000000000
-  },
-  {
-    address = "TLLM21wteSPs4hKjbxgmH1L6poyMjeTbHm"
-    owner = "TS1hu4ZCcwBFYpQqUGoWy1GWBzamqxiT5W"
-  },
-  {
-    address = "TRY18iTFy6p8yhWiCt1dhd2gz2c15ungq3"
-    trc10Id = "1000001"
-    trc10Balance = 100000000
-  },
-  {
-    address = "TKmyxLsRR2FWMVEHaQA2pZh1xB7oXPXzG1"
-  }
-]
-
-trc20Contracts = [
-  {
-    contractAddress = "TR7NHqjeKQxGTCi8q8ZY4pL8otSzgjLj6t"
-    balancesSlotPosition = 0
-    address = "TRY18iTFy6p8yhWiCt1dhd2gz2c15ungq3"
-    balance = "98800000000000000"
-  },
-  {
-    contractAddress = "TSSMHYeV2uE9qYH95DqyoCuNCzEL1NvU3S"
-    balancesSlotPosition = 0
-    address = "TRY18iTFy6p8yhWiCt1dhd2gz2c15ungq3"
-    balance = "128745186062510400000000000"
-  }
-]
-
-latestBlockHeaderTimestamp = 1739269661000
-maintenanceTimeInterval = 21600000
-nextMaintenanceTime = 1735628894000
