+++ tools/gradlew/settings.gradle
@@ -1,7 +1,5 @@
 rootProject.name = 'tron-docker'
 include('docker')
 project(':docker').projectDir = new File(rootDir, '../docker')
-include('dbfork')
-project(':dbfork').projectDir = new File(rootDir, '../dbfork')
 include('toolkit')
 project(':toolkit').projectDir = new File(rootDir, '../toolkit')
