+++ tools/toolkit/src/test/java/org/tron/plugins/DbLiteRocksDbTest.java
@@ -1,7 +1,6 @@
 package org.tron.plugins;
 
 import java.io.IOException;
-import org.junit.Ignore;
 import org.junit.Test;
 
 public class DbLiteRocksDbTest extends DbLiteTest {
